Cleanup (#3)
* Cleanup

* More cleanup

* Even more cleanup

* ...

* Fix missing header

* Allow digitset with base = 1

+++ test/test_introspection.cpp
@@ -187,7 +187,7 @@ TEST(Introspection, StaticForAggregate)
   auto fun = cbr::overload
   {
     [&](const std::string & v) {
-      i += v.size();
+      i += static_cast<int>(v.size());
     },
     [&](int & v) {
       i += v;
